+++ pkg/bgp/mp-nlri.go
@@ -5,6 +5,7 @@ import (
 	"github.com/sbezverk/gobmp/pkg/evpn"
 	"github.com/sbezverk/gobmp/pkg/flowspec"
 	"github.com/sbezverk/gobmp/pkg/ls"
+	"github.com/sbezverk/gobmp/pkg/mcastvpn"
 	"github.com/sbezverk/gobmp/pkg/rtc"
 	"github.com/sbezverk/gobmp/pkg/srpolicy"
 )
@@ -21,6 +22,7 @@ type MPNLRI interface {
 	GetNLRI73() (*srpolicy.NLRI73, error)
 	GetFlowspecNLRI() (*flowspec.NLRI, error)
 	GetNLRIRTC() (*rtc.Route, error)
+	GetNLRIMCASTVPN() (*mcastvpn.Route, error)
 	GetNextHop() string
 	IsIPv6NLRI() bool
 	IsNextHopIPv6() bool
@@ -87,6 +89,12 @@ func NLRIMessageType(afi uint16, safi uint8) int {
 		// AFI 2 and SAFI 132 Route Target Constraint v6
 	case afi == 2 && safi == 132:
 		return 31
+		// AFI 1 and SAFI 5 MCAST-VPN v4
+	case afi == 1 && safi == 5:
+		return 32
+		// AFI 2 and SAFI 5 MCAST-VPN v6
+	case afi == 2 && safi == 5:
+		return 33
 	}
 
 	return 0