+++ pkg/message/process-mp-update.go
@@ -177,6 +177,24 @@ func (p *producer) processMPUpdate(nlri bgp.MPNLRI, operation int, ph *bmp.PerPe
 				return
 			}
 		}
+	case 32:
+		fallthrough
+	case 33:
+		msgs, err := p.mcastvpn(nlri, operation, ph, update)
+		if err != nil {
+			glog.Errorf("failed to produce mcastvpn messages with error: %+v", err)
+			return
+		}
+		for _, m := range msgs {
+			topicType := bmp.MCASTVPNV6Msg
+			if m.IsIPv4 {
+				topicType = bmp.MCASTVPNV4Msg
+			}
+			if err := p.marshalAndPublish(&m, topicType, []byte(m.RouterHash), false); err != nil {
+				glog.Errorf("failed to process MCAST-VPN message with error: %+v", err)
+				return
+			}
+		}
 	case 71:
 		p.processNLRI71SubTypes(nlri, operation, ph, update)
 	}